+++ main_ewm/main_ewm_service/src/main/java/ru/practicum/controller/ControllerUser.java
@@ -4,10 +4,10 @@ import events.dto.EventFullDto;
 import events.dto.EventShortDto;
 import events.dto.NewEventDto;
 import events.dto.ParticipationRequestDto;
-import events.enum_events.StatusUserRequestEvent;
-import events.model.EventRequestStatusUpdateRequest;
-import events.model.EventRequestStatusUpdateResult;
-import events.model.UpdateEventUserRequest;
+import ru.practicum.services.events.enum_events.StatusUserRequestEvent;
+import ru.practicum.services.events.model.EventRequestStatusUpdateRequest;
+import ru.practicum.services.events.model.EventRequestStatusUpdateResult;
+import ru.practicum.services.events.model.UpdateEventUserRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
