+++ main_ewm/main_ewm_service/src/main/java/ru/practicum/services/events/enum_events/StatusAction.java
@@ -1,4 +1,4 @@
-package events.enum_events;
+package ru.practicum.services.events.enum_events;
 
 public enum StatusAction {
     PUBLISH_EVENT,
