+++ main_ewm/main_ewm_service/src/main/java/ru/practicum/services/events/enum_events/StatusReview.java
@@ -1,4 +1,4 @@
-package events.enum_events;
+package ru.practicum.services.events.enum_events;
 
 public enum StatusReview {
     SEND_TO_REVIEW,
