+++ main_ewm/main_ewm_service/src/main/java/ru/practicum/services/events/model/EventRequestStatusUpdateResult.java
@@ -1,4 +1,4 @@
-package events.model;
+package ru.practicum.services.events.model;
 
 import events.dto.ParticipationRequestDto;
 import lombok.AllArgsConstructor;
