+++ main_ewm/main_ewm_service/src/main/java/ru/practicum/services/events/model/StatusEvent.java
@@ -1,4 +1,4 @@
-package events.enum_events;
+package ru.practicum.services.events.model;
 
 public enum StatusEvent {
     PENDING,
