+++ main_ewm/main_ewm_service/src/main/java/ru/practicum/services/events/model/UpdateEventAdminRequest.java
@@ -1,14 +1,13 @@
-package events.model;
+package ru.practicum.services.events.model;
 
-import categories.dto.CategoryDto;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import events.enum_events.StatusReview;
 import events.model.LocationEvent;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import ru.practicum.services.events.enum_events.StatusAction;
 
 import javax.validation.constraints.Size;
 import java.time.LocalDateTime;
@@ -17,11 +16,12 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Getter
 @Setter
-public class UpdateEventUserRequest {
+public class UpdateEventAdminRequest {
+
     @Size(min = 20, max = 2000)
     private String annotation;
 
-    private CategoryDto category;
+    private Long category;
 
     @Size(min = 20, max = 7000)
     private String description;
@@ -34,13 +34,14 @@ public class UpdateEventUserRequest {
     @JsonProperty(value = "paid")
     private Boolean paid;
 
-    private Integer participantLimit;
+    private Long participantLimit;
 
     @JsonProperty(value = "requestModeration")
     private Boolean requestModeration;
 
-    private StatusReview stateAction;
+    private StatusAction stateAction;
 
     @Size(min = 3, max = 120)
     private String title;
+
 }
