+++ main_ewm/main_ewm_service/src/main/java/ru/practicum/services/events/model/UpdateEventUserRequest.java
@@ -1,12 +1,14 @@
-package events.model;
+package ru.practicum.services.events.model;
 
+import categories.dto.CategoryDto;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import events.enum_events.StatusAction;
+import events.model.LocationEvent;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import ru.practicum.services.events.enum_events.StatusReview;
 
 import javax.validation.constraints.Size;
 import java.time.LocalDateTime;
@@ -15,12 +17,11 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Getter
 @Setter
-public class UpdateEventAdminRequest {
-
+public class UpdateEventUserRequest {
     @Size(min = 20, max = 2000)
     private String annotation;
 
-    private Long category;
+    private CategoryDto category;
 
     @Size(min = 20, max = 7000)
     private String description;
@@ -33,14 +34,13 @@ public class UpdateEventAdminRequest {
     @JsonProperty(value = "paid")
     private Boolean paid;
 
-    private Long participantLimit;
+    private Integer participantLimit;
 
     @JsonProperty(value = "requestModeration")
     private Boolean requestModeration;
 
-    private StatusAction stateAction;
+    private StatusReview stateAction;
 
     @Size(min = 3, max = 120)
     private String title;
-
 }
