+++ main_ewm/main_ewm_service/src/main/java/ru/practicum/services/participants_request/mapper/MapperParticipationRequest.java
@@ -14,7 +14,7 @@ public class MapperParticipationRequest {
                 request.getEvent().getId(),
                 request.getId(),
                 request.getRequester().getId(),
-                request.getStatus());
+                request.getStatus().toString());
     }
 
     public static List<ParticipationRequestDto> toDto(Iterable<ParticipationRequest> requests) {
